Task/upgrade dotnet 10 (#9)
* Updates Stripe Android and iOS Bindings

Updates the Stripe Bindings to the latest versions.
This includes updating the Android SDK version, Kotlin standard libraries,
AndroidX libraries, and adjusts metadata transforms.

* Updates target frameworks to net10.0

Updates the target frameworks for Android and iOS projects to .NET 10.0.

This ensures compatibility with the latest .NET runtime and allows the projects to take advantage of new features and improvements.

---------

Co-authored-by: Ricardo Bárbara <ricardo.barbara@ipn.pt>